+++ src/components/icon/__spec__.js
@@ -71,6 +71,14 @@ describe('Icon', () => {
       expect(svg.getAttribute('class')).toEqual("ui-icon__svg ui-icon__svg--basket");
     });
 
+    it('calls the render newtab icon method', () => {
+      instance = TestUtils.renderIntoDocument(<Icon type='newtab' />);
+      instance.renderIcon;
+      span = TestUtils.findRenderedDOMComponentWithTag(instance, 'span');
+      svg = span.children[0];
+      expect(svg.getAttribute('class')).toEqual("ui-icon__svg ui-icon__svg--newtab");
+    });
+
     it('calls the render phone icon method', () => {
       instance = TestUtils.renderIntoDocument(<Icon type='phone' />);
       instance.renderIcon;